% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assign_prob_and_strata}
\alias{assign_prob_and_strata}
\title{Assign Probability and Strata according to Reference Table}
\usage{
assign_prob_and_strata(data, Z, Z_table, probabilities, merge_strata = TRUE)
}
\arguments{
\item{data}{(\code{data.frame}) Input data frame.}

\item{Z}{(\code{vector}) The variable names in \code{data} and \code{Z_table} that define \code{stratum}.}

\item{Z_table}{(\code{data.frame}) A reference table where rows define strata and columns include \code{Z} and \code{probabilities}.}

\item{probabilities}{(\code{vector}) The names of columns in \code{Z_table} representing treatment assignment probabilities.}

\item{merge_strata}{(\code{logical}) Whether to merge strata with identical probabilities.}
}
\description{
Assign Probability and Strata according to Reference Table
}
\details{
The output column name is \code{stratum}.
}
\examples{
Z <- c("t", "subtype")
probabilities <- c("trt.1", "trt.2", "trt.3", "trt.4")
Z_table <- unique(example[c(Z, probabilities)])
new_example = assign_prob_and_strata(
                            data = example[Z],
                            Z = Z,
                            Z_table =  Z_table,
                            probabilities = probabilities,
                            merge_strata = TRUE
                            )
head(new_example)

}
