% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_wt.R
\name{estimation_wt}
\alias{estimation_wt}
\title{Compute Estimates and Covariance Matrix}
\usage{
estimation_wt(
  ret,
  y,
  treatment,
  treatments_for_compare,
  data,
  prob_mat,
  stabilize = T
)
}
\arguments{
\item{ret}{counterfactual prediction}

\item{y}{Observed outcome}

\item{treatment}{name of treatment}

\item{treatments_for_compare}{description}

\item{data}{(\code{data.frame}) data}

\item{prob_mat}{(\code{data.frame}) treatment assignment probability}

\item{stabilize}{(\code{logical}) whether to stabilize}
}
\description{
Compute Estimates and Covariance Matrix
}
