% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Trial Data}
\format{
A data frame with 500 rows and 8 columns:
\describe{
\item{xc, xb, subtype}{The covariates.}
\item{t, substudy}{The enrollment window and substudy assignment}
\item{treatment}{The treatment assignment, "1", "2", "3" and "4"}
\item{y}{The continuous response.}
\item{y_b}{The binary response, I(y>3).}
\item{s12, s12.2, s12.error, s13, s14}{The stratification variables.}
}
}
\source{
The data is generated by Yuhan.
}
\usage{
example
}
\description{
This dataset is generated based on the first three enrollment windows outlined in Figure 1(a) in our paper.
}
\keyword{dataset}
