% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_counterfactual.R
\name{predict_counterfactual}
\alias{predict_counterfactual}
\title{Counterfactual Prediction}
\usage{
predict_counterfactual(
  fit,
  treatment,
  treatments_for_compare,
  prob_mat,
  data,
  ...
)
}
\arguments{
\item{fit}{fitted object.}

\item{treatment}{name of treatment column}

\item{treatments_for_compare}{(\code{character}) Treatments for comparison}

\item{data}{(\code{data.frame}) raw dataset.}
}
\value{
Numeric matrix of counter factual prediction.
}
\description{
Obtain counterfactual prediction of a fit.
}
