% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prob_mat}
\alias{prob_mat}
\title{Treatment Assignment Probability Data}
\format{
A data frame with 500 rows and 4 columns:
\describe{
\item{1, 2, 3, 4}{The assgiment probabilities of trt 1, 2, 3, and 4.}
}
}
\source{
The data is generated by Yuhan.
}
\usage{
prob_mat
}
\description{
This dataset gives the assignment probabilities in the example dataset.
}
\keyword{dataset}
