% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_wt.R
\name{robin_wt}
\alias{robin_wt}
\title{Inverse probability weighting based model}
\usage{
robin_wt(
  formula,
  data,
  treatment,
  prob_mat,
  treatments_for_compare,
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = T,
  ...
)
}
\arguments{
\item{formula}{(\code{formula}) A formula of analysis.}

\item{data}{(\code{data.frame}) Input data frame.}

\item{treatment}{(\code{character(1)}) A string name of treatment assignment.}

\item{prob_mat}{(\code{data.frame}) Treatment assignment probability matrix}

\item{treatments_for_compare}{(\code{vector}) Treatments for comparison}

\item{contrast}{(\code{function} or \code{character(1)}) A function to calculate the treatment effect, or character of
"difference", "risk_ratio", "odds_ratio" for default contrasts.}

\item{contrast_jac}{(\code{function}) A function to calculate the Jacobian of the contrast function. Ignored if using
default contrasts.}

\item{family}{(\code{family}) A family object of the glm model.}

\item{stabilize}{(\code{logical}) Whether to stabilize}
}
\description{
Inverse probability weighting based model
}
\examples{
robin_wt(
  formula = y ~ treatment * x,
  data = dummy_data,
  treatment = treatment,
  prob_mat = prob_mat,
  treatments_for_compare = c(1,2),
  contrast = "difference",
  contrast_jac=NULL,
  family=gaussian(),
  stabilize=T)
}
