% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobinCID-package.R
\docType{package}
\name{RobinCID-package}
\alias{RobinCID-package}
\alias{RobinCID}
\title{\code{RobinCID} Package}
\description{
\code{RobinCID} implements unbiased prediction and robust inference in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Eureeca/RobinCID}
  \item Report bugs at \url{https://github.com/Eureeca/RobinCID/issues}
}

}
\author{
\strong{Maintainer}: Yuhan Qian \email{yhqian@uw.edu} [copyright holder]

Authors:
\itemize{
  \item Ting Ye
  \item Yanyao Yi
  \item Marlena Bannick
}

}
