% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_effect.R
\name{estimate_effect}
\alias{estimate_effect}
\title{Compute Estimates and Covariance Matrix}
\usage{
estimate_effect(
  ret,
  y,
  treatment,
  treatments_for_compare,
  data,
  prob_mat,
  stratification,
  stabilize
)
}
\arguments{
\item{ret}{counterfactual prediction}

\item{y}{Observed outcome}

\item{treatment}{name of treatment}

\item{treatments_for_compare}{description}

\item{data}{(\code{data.frame}) data}

\item{prob_mat}{(\code{data.frame}) treatment assignment probability}

\item{stratification}{(\code{character}) name of stratification variable}

\item{stabilize}{(\code{logical}) whether to stabilize}
}
\description{
Compute Estimates and Covariance Matrix
}
