% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_estimate.R
\name{robin_estimate}
\alias{robin_estimate}
\title{Robust inference method}
\usage{
robin_estimate(
  formula,
  data,
  treatment,
  prob_mat = NULL,
  stratification = NULL,
  treatments_for_compare,
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = TRUE,
  alpha = 0.05,
  method = NULL,
  ...
)
}
\arguments{
\item{formula}{(\code{formula}) A formula of analysis.}

\item{data}{(\code{data.frame}) Input data frame.}

\item{treatment}{(\code{character}) A string name of treatment assignment.}

\item{prob_mat}{(\code{data.frame}) Treatment assignment probability matrix}

\item{stratification}{(\code{character}) A string name of stratification variable. Default: \code{NULL}}

\item{treatments_for_compare}{(\code{vector}) Treatments for comparison}

\item{contrast}{(\code{function} or \code{character}) A function to calculate the treatment effect, or character of
"difference", "risk_ratio", "odds_ratio" for default contrasts.}

\item{contrast_jac}{(\code{function}) A function to calculate the Jacobian of the contrast function. Ignored if using
default contrasts.}

\item{family}{(\code{family}) A family object of the glm model. Default: \code{gaussian()}.}

\item{stabilize}{(\code{logical}) Whether to stabilize. Default: TRUE.}

\item{alpha}{(\code{double}) Nominal level. Default: 0.05.}

\item{method}{Estimation method. Either "wt" or "ps".}

\item{...}{Additional arguments passed to \code{glm}}
}
\description{
Provides robust inference methods via either weighting or post stratification.
}
\details{
If family is \code{MASS::negative.binomial(NA)}, the function will use \code{MASS::glm.nb} instead of \code{glm}.
}
