% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_estimate.R
\name{robin_ps}
\alias{robin_ps}
\title{Post-Stratification Based Inference}
\usage{
robin_ps(
  formula,
  data,
  treatment,
  probabilities = NULL,
  post_strata = NULL,
  treatments_for_compare,
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  alpha = 0.05,
  ...
)
}
\arguments{
\item{formula}{(\code{formula}) A formula of analysis.}

\item{data}{(\code{data.frame}) Input data frame.}

\item{treatment}{(\code{character}) A string name of treatment assignment.}

\item{probabilities}{(\code{vector}) A character vector specifying column names of treatment assignment probabilities in the \code{data}.
The column names must represent the treatment levels in the study and should include at least the treatments specified in the \code{treatments_for_compare}.}

\item{post_strata}{(\code{character}) A string name of post-stratification variable. Default: \code{NULL}}

\item{treatments_for_compare}{(\code{vector}) A character vector specifying exactly two treatment levels to compare.
The specified treatments must be present in both the \code{treatment} variable of the \code{data} and the column names of the \code{probability}.}

\item{contrast}{(\code{function} or \code{character}) A function to calculate the treatment effect, or character of
"difference", "risk_ratio", "odds_ratio" for default contrasts.}

\item{contrast_jac}{(\code{function}) A function to calculate the Jacobian of the contrast function. Ignored if using
default contrasts.}

\item{family}{(\code{family}) A family object of the glm model. Default: \code{gaussian()}.}

\item{alpha}{(\code{double}) Nominal level. Default: 0.05.}

\item{...}{Additional arguments passed to \code{glm}.}
}
\description{
Provides robust inference methods via post stratification.
}
\details{
If family is \code{MASS::negative.binomial(NA)}, the function will use \code{MASS::glm.nb} instead of \code{glm}.
}
\examples{
robin_ps(
  formula = y ~ xb + xc,
  data = example,
  treatment = "treatment",
  probabilities = c("trt.1", "trt.2", "trt.3", "trt.4"),
  post_strata = NULL,
  treatments_for_compare = c("trt.1", "trt.2"),
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  alpha = 0.05
)
}
