% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_estimate.R
\name{robin_wt}
\alias{robin_wt}
\title{Inverse Probability Weighting Based Inference}
\usage{
robin_wt(
  formula,
  data,
  treatment,
  prob_mat,
  treatments_for_compare,
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = TRUE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{formula}{(\code{formula}) A formula of analysis.}

\item{data}{(\code{data.frame}) Input data frame.}

\item{treatment}{(\code{character}) A string name of treatment assignment.}

\item{prob_mat}{(\code{data.frame}) A data frame containing the treatment assignment probabilities.
Each row corresponds to an observation in the \code{data}.
The column names must represent the treatment levels in the study and should include at least the treatments specified in the \code{treatments_for_compare}.}

\item{treatments_for_compare}{(\code{vector}) A character vector specifying exactly two treatment levels to compare.
The specified treatments must be present in both the \code{treatment} variable of the \code{data} and the column names of the \code{prob_mat}.}

\item{contrast}{(\code{function} or \code{character}) A function to calculate the treatment effect, or character of
"difference", "risk_ratio", "odds_ratio" for default contrasts.}

\item{contrast_jac}{(\code{function}) A function to calculate the Jacobian of the contrast function. Ignored if using
default contrasts.}

\item{family}{(\code{family}) A family object of the glm model. Default: \code{gaussian()}.}

\item{stabilize}{(\code{logical}) Whether to stabilize. Default: TRUE.}

\item{alpha}{(\code{double}) Nominal level. Default: 0.05.}

\item{...}{Additional arguments passed to \code{glm}.}
}
\description{
Provides robust inference methods via inverse probability weighting.
}
\details{
If family is \code{MASS::negative.binomial(NA)}, the function will use \code{MASS::glm.nb} instead of \code{glm}.
}
\examples{
robin_wt(
  formula = y ~ xb + xc,
  data = example,
  treatment = "treatment",
  prob_mat = prob_mat,
  treatments_for_compare = c("1", "2"),
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = TRUE,
  alpha = 0.05
)
}
