% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_estimate.R
\name{robin_wt}
\alias{robin_wt}
\title{Inverse Probability Weighting Based Inference}
\usage{
robin_wt(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{glm}}
}
\value{
See \code{robin_estimate} for details.
}
\description{
A wrapper function for \code{robin_estimate} with the method set to "wt".
}
\examples{
robin_wt(
  formula = y ~ xb + xc,
  data = example,
  treatment = "treatment",
  prob_mat = prob_mat,
  treatments_for_compare = c("1", "2"),
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = TRUE,
  alpha = 0.05
)
}
