% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_estimate.R
\name{robin_wt}
\alias{robin_wt}
\title{Inverse Probability Weighting Based Inference}
\usage{
robin_wt(
  formula,
  data,
  treatment,
  probabilities,
  Z = NULL,
  Z_table = NULL,
  treatments_for_compare,
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = TRUE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{formula}{(\code{formula}) A formula of analysis.}

\item{data}{(\code{data.frame}) Input data frame.}

\item{treatment}{(\code{character}) A string name of treatment assignment.}

\item{probabilities}{(\code{vector}) A character vector specifying column names of treatment assignment probabilities in the \code{data}.
The column names must represent the treatment levels in the study and should include at least the treatments specified in the \code{treatments_for_compare}.}

\item{Z}{(\code{vector}) The variable names in \code{data} and \code{Z_table} that define \code{stratum}.}

\item{Z_table}{(\code{data.frame}) A reference table where rows define strata and columns include \code{Z} and \code{probabilities}.}

\item{treatments_for_compare}{(\code{vector}) A character vector specifying exactly two treatment levels to compare.
The specified treatments must be present in both the \code{treatment} variable of the \code{data} and the column names of the \code{probability}.}

\item{contrast}{(\code{function} or \code{character}) A function to calculate the treatment effect, or character of
"difference", "risk_ratio", "odds_ratio" for default contrasts.}

\item{contrast_jac}{(\code{function}) A function to calculate the Jacobian of the contrast function. Ignored if using
default contrasts.}

\item{family}{(\code{family}) A family object of the glm model. Default: \code{gaussian()}.}

\item{stabilize}{(\code{logical}) Whether to stabilize. Default: TRUE.}

\item{alpha}{(\code{double}) Nominal level. Default: 0.05.}

\item{...}{Additional arguments passed to \code{glm}.}
}
\description{
Provides robust inference methods via inverse probability weighting.
}
\details{
If family is \code{MASS::negative.binomial(NA)}, the function will use \code{MASS::glm.nb} instead of \code{glm}.
}
\examples{
probabilities <- c("trt.1", "trt.2", "trt.3", "trt.4")

robin_wt(
  formula = y ~ xb + xc,
  data = example,
  treatment = "treatment",
  probabilities = probabilities,
  treatments_for_compare = c("trt.1", "trt.2"),
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = TRUE,
  alpha = 0.05
)

Z <- c("t", "subtype")
Z_table <- unique(example[c(Z, probabilities)])
robin_wt(
  formula = y ~ xb + xc,
  data = example[setdiff(names(example), probabilities)],
  treatment = "treatment",
  probabilities = probabilities,
  Z = Z,
  Z_table = Z_table,
  treatments_for_compare = c("trt.1", "trt.2"),
  contrast = "difference",
  contrast_jac = NULL,
  family = gaussian(),
  stabilize = TRUE,
  alpha = 0.05
)
}
